src/modules/categories/useCases/ListProductsByCategory/ListProductsByCategoryUseCase.ts
import { inject, injectable } from 'tsyringe';

import { Category } from '@modules/categories/infra/entities/Category';
import { ICategoriesRepository } from '@modules/categories/repositories/ICategoriesRepository';
import { AppError } from '@shared/errors/AppError';

@injectable()
export class ListProductsByCategoryUseCase {
  constructor(
    @inject('CategoriesRepository')
    private categoriesRepository: ICategoriesRepository
  ) {}

  async execute(id: string): Promise<Category> {
    const category = await this.categoriesRepository.findById(id);

    if (!category) {
      throw new AppError('Product not found');
    }

    return category;
  }
}
